requirements.txt
Flask==2.3.2
keras==2.12.0
numpy==1.24.3
Pillow==10.0.0